adding next toploader

+++ src/app/api/clubs/[clubId]/requests/decline-request/route.ts
@@ -1,5 +1,6 @@
 
 
+import { validateRequest } from "@/server/auth";
 import { db } from "@/server/db"
 import { NextRequest, NextResponse } from "next/server"
 
@@ -9,6 +10,16 @@ type Payload = {
 export async function POST(req: NextRequest) {
     const { id }:Payload = await req.json()
     try {
+        const user = await validateRequest();
+        if (!user?.isMentor || user?.role === "ADMIN") {
+
+            return NextResponse.json({
+                status: 403,
+                message: "You're not supposed to decline"
+
+            })
+        }
+
         await db.request.delete({
             where: {
                 id
